Flying Enemy and Ally Patrol
Flying enemy exists and has the same behaviour as the other enemies. It also has a simple animation. Might need some tuning and have to explore if there is any bugs. Ally Patrol also exists now, but it is very in the early stages of development. Have to work on it.