Messed around with NavMesh and Messed around with Towers
Messed around with the NavMesh to make it a bit more smoother. It can be more smooth than it is now, but I have yet to find a way to do it. Firework Tower Behaviour was made and I added a different version of it to the upgrades. Same with the Kunai Tower except I only had to add it to the upgrades.