Overhauled EnemySpawner and Round Count
Added a UI that counts the round. It is pretty garbage, so change it if you want later on or make it better. I completely overhauled the enemyspawner and now there is only one spawner for the whole scene. Enemies spawn based on a percentage depending on the total amount of enemies that have to spawn and the round number. Noticed that the game crashes the more coins you have or something else is causing it. I don't think it is my code, but I copy pasted the original code for testing later.