Added RigidB to Colliders & Bamboo
I added RigidBodies to all bamboo clusters, colliders, and one of the navmesh obstacles inside the mountain. I made them have the maximum amount of mass possible and also made them kinematic and not use gravity. Couldn't test because Gamble is currently broken and I also saw that Gamble had two rigidbodies for some reason. One of them is only 1 in mass and the other has the maximum amount.