Bull Behaviour + Pathing
I made a new BullBehaviour for the new Bull enemy prefab. The pathing is inside this behaviour and it uses the position of gameobjects inside the pathing array. Also added a new tag for them. It should work fine, but I haven't tested it much when the game runs. It also needs a spawner.