Bull Spawner
The bull now spawns in round 9. It is in the EnemySpawner script. I basically made it so that it decreases the remainingEnemies in the RoundManager to 1 then once the Bull dies it just makes it 0. I added an extra line of code in the BullBehaviour for it to decrease the remainingEnemies once it dies.