EnemyTower + Minor Bug Fixes
The EnemyTower script has been created. This keeps track of the health and also despawns enemies that are still alive. Used the EnemySpawner for it to only spawn in front of the tower. I also halved the spawn interval when round 10 hits. I edited the code for all enemies to stop decreasing remainingEnemies in the RoundManager once round 10 hits as well.  Also some minor bug fixes like the Dojo not having their spawn point assigned and allies not detecting spiders since the SpiderBehaviour was never referenced in their code.  Also, the wall towers had some colliders enabled when they shouldn't.