NavMesh + Pathing
I made pathings for enemies. I didn't use the spawners to do it, but instead I added a lot of checks for conditions using colliders and then setting a path (I have a headache). Apart from this, I added the navmesh to the new map and made sure the round system works. I have no clue what happens at around 9 when the bull is supposed to spawn, but that is a problem for another day.