NavMesh + Bug Fixes
Added the navmesh to the new map. I also fixed an issue with the look direction. I added an invisible plane around where the character walks. It has a trigger collider and shouldn't affect anything. Either way, if it brings trouble I can try to fix it actually doing coding instead of this lazy fix.