Archer && Brute
Added the Brute to the spawner and made it work. There is a bug with it currently that it instakills anything in the vicinity. I am assuming that is due to the animator and can't do much about it. Made the archers works too. They spawn and go around the map. Their perimeter basically covers 3/4ths of the whole map. I also made an adjustment to the ProjectileBehaviour and got rid of the hitting the ground condition to get destroyed. Did this cause of the invisible ground plane for the look direction.