New Level Pathing && Spawning
I was able to program the pathing and it looks fine. They endlessly loop around the level and are close enough to attack the defense point. As for the spawning I barely coded anything and just added a condition for faction spawns. There is basically 2 spawners in the scene instead of the usual 1.